requirements.txt
grpcio==1.42.0
grpcio-tools==1.42.0
torch
matplotlib
torchinfo
PyGetWindow
pyobjc-framework-Quartz        # M1 mac only?
pyscreenshot
opencv-python==3.4.15.55
Pillow
# install hydra using: pip install hydra-core --upgrade
